+++ src/webhooks/index.ts
@@ -1,20 +1,32 @@
 import { createHmac } from "crypto";
 import type { Paddle } from "../types";
+import type { PaddleAPI } from "../api/types";
 
 /**
- * Validates Paddle event using the secrete key and the signature.
+ * Validates Paddle event using the secret key and the signature.
  *
+ * @param client - The Paddle Billing client
  * @param secret - The secret key
  * @param signature - The Paddle signature
  * @param body - The raw body string to validate
  *
  * @returns true if the body is authentic
  */
-export function parseWebhookBody(
+export function parseWebhookBody<DataDef extends PaddleAPI.CustomDataDef>(
+  _client: PaddleAPI.Client<DataDef> | null,
   secret: string,
   signature: string,
   body: string
-): Paddle.Event | null {
+): Paddle.Event<
+  PaddleAPI.CustomData<DataDef["Price"]>,
+  PaddleAPI.CustomData<DataDef["Product"]>,
+  PaddleAPI.CustomData<DataDef["SubscriptionItem"]>,
+  PaddleAPI.CustomData<DataDef["Subscription"]>,
+  PaddleAPI.CustomData<DataDef["Transaction"]>,
+  PaddleAPI.CustomData<DataDef["Customer"]>,
+  PaddleAPI.CustomData<DataDef["Address"]>,
+  PaddleAPI.CustomData<DataDef["Business"]>
+> | null {
   const captures = signature.match(/^ts=(\d+);h1=(.+)$/);
   if (!captures) return null;